requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
wandb>=0.15.0
PyYAML>=6.0
